+++ resources/views/components/livewire/loading.blade.php
@@ -1,4 +1,4 @@
-<div class="flex flex-row justify-center items-center w-full h-full fixed select-none z-50" wire:loading.flex
+<div class="flex flex-row justify-center items-center w-full h-full fixed select-none z-[300]" wire:loading.flex
      style="left: 50%;top: 50%;transform: translateX(-50%) translateY(-50%);background: rgba(255,255,255,0.7);">
     <i class="fa-solid fa-spinner fa-spin-pulse text-[5em]"></i>
     <div class="ml-10 text-primary text-xl">