+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - add development page with different details, and a config to enable (public) access to it
 - add possibility to update page title
+### Changed
+- input switch works now with events
 
 ## v0.7.6 - 2024-04-08
 ### Added